Исправления показа полноэкранного изображения
