+++ js/renderPicture.js
@@ -1,3 +1,5 @@
+import {openFullPictureWindow} from './openFullPictureWindow.js';
+
 const pictureList = document.querySelector('.pictures');
 const pictureTemplate = document.querySelector('#picture').content.querySelector('.picture');
 
@@ -9,6 +11,11 @@ const renderPictures = (pictures) => {
     picture.querySelector('.picture__img').src = photo.url;
     picture.querySelector('.picture__comments').textContent = photo.comments.length;
     picture.querySelector('.picture__likes').textContent = photo.likes;
+    picture.addEventListener('click', (evt) => {
+      evt.preventDefault();
+      openFullPictureWindow(photo);
+    });
+
     pictureListFragment.appendChild(picture);
   });
